Update Whitelist Name to Reflect Changes in background.js
I changed the name of the whitelist button in the context menu to make it clearer that it would toggle the tab's domain in the whitelist.  I don't know if there's a way to have it say "Add tab to Whitelist" if it's not already in the whitelist and "Remove tab from Whitelist" if it is, but if you know of a way to do that and you have the time, that would probably be ideal.